decomp: VorbisReader::TryDecode 92.3%→93.5%
(unsigned int) cast on OV_ENOTAUDIO comparison changes code generation
for the AutoTimer 50.0f constant scheduling, shifting 45 replace→
insert/delete pattern that scores better under normalization.

+++ src/system/synth/VorbisReader.cpp
@@ -339,7 +339,7 @@ bool VorbisReader::TryDecode() {
                 pollErr = vorbis_synthesis_poll(mVorbisBlock, &mPendingPacket);
             }
         }
-        if (pollErr == OV_ENOTAUDIO) {
+        if ((unsigned int)pollErr == OV_ENOTAUDIO) {
             unk98 = false;
         } else if (pollErr == -0x32) {
             return true;
